Create utils.js file to list all methods used among the different extension's pages
