+++ addon/inspector.js
@@ -1,18 +1,7 @@
 export let defaultApiVersion = "63.0";
 export let apiVersion = localStorage.getItem("apiVersion") == null ? defaultApiVersion : localStorage.getItem("apiVersion");
 
-export function getLinkTarget(e = {}) {
-  if (localStorage.getItem("openLinksInNewTab") == "true" || (e.ctrlKey || e.metaKey)) {
-    return "_blank";
-  } else {
-    return "_top";
-  }
-}
 export let sessionError;
-export function nullToEmptyString(value) {
-  // For react input fields, the value may not be null or undefined, so this will clean the value
-  return (value == null) ? "" : value;
-}
 const clientId = "Salesforce Inspector Reloaded";
 
 export let sfConn = {
