+++ addon/utils.js
@@ -1,0 +1,12 @@
+export function getLinkTarget(e = {}) {
+  if (localStorage.getItem("openLinksInNewTab") == "true" || (e.ctrlKey || e.metaKey)) {
+    return "_blank";
+  } else {
+    return "_top";
+  }
+}
+
+export function nullToEmptyString(value) {
+  // For react input fields, the value may not be null or undefined, so this will clean the value
+  return (value == null) ? "" : value;
+}
